Updated Time Scheduling Code
* Added a Time Scheduling Tolerance which fits into each sensors time
slot. All measurements MUST begin in this tolerance phase, the rest of
the time slot is reserved for existing measurements (this stops
ultrasonic pulses from leaking into the next time slot). Time Scheduling
Tolerance can be specified in the general INI file with
"SENSOR_CLIENT_TIME_SCHEDULING_TOLERANCE" and is recorded in ms as a
float.
* Fixed critical bug that allowed polling frequencies 100x higher than
what is physically possible (didn't convert cm --> m)
* Changed several variable names throughout the solution to specify what
unit they are measured in (e.g. VariableName_ms)
* Renamed SensorClient.IScheduler.SendData() to Start() to better
reflect the asynchronous intention of the interface.