InitialEstimator averages, Display draw additional
* The InitialEstimator now averages all candidates in an anchors
proximity rather than only choosing the first candidates position
* The DisplayServer now draws additional state information as yellow
dots (currently used to show declined candidates). The SensorServer now
correspondingly keeps tracks of and sends this information.
* The IEstimator interface now requires estimators to store the current
stage of calculations for candidates and additional unaccepted
candidates in corresponding properties, which may be discarded after a
new calculation
* Fixed bug with trilateration calculations, arctan was used incorrectly
with assumption of positive-only results.
* WSN.Variables now has proper support for returning bool types (no need
to convert to string first)
* INI File changes:
* Added: TRILATERATOR_NOISELESS_0D_DISTANCE_TOLERANCE : double that
denotes the tolerance the InitialEstimator should use for determining if
candidates belong to an anchor
* Added: DRAW_ADDITONAL : bool, determines if the newly supported
additional state should be drawn (the yellow circles)
* Added: DRAW_XY_AXIS_1_TO_1 : bool, determines if the DisplayServer
should set the length of the X and Y axes to be equivalent. Setting this
to true will prevent the previous problem of squishing and stretching of
images, but may cause poor use of screen space in certain extreme cases.