KalmanFilter2D + Fix to previous commit
Implemented a two dimension Kalman Filter on the sensor server, which
makes use of Matlab. It is currently not in use, it will be implemented
in InitialEstimator.cs when the track association problem is solved.

Creating a new KalmanFilter2D object can take a very long time (>3
seconds), so when it is finally implemented it may be worth initialising
it early on, before the sensor server / display server start
communicating.

My previous commit on "InitialEstimator averages..." looks like it may
have missed some changes, they're included here.