Added Real State Parsing
Added a Real State parser class that parses a .txt./.csv hybrid file
which describes a real state, for comparison with the estimated object
position(s).

Currently only one real object is allowed, though the change to permit
multiple would be trivial.

Currently a real state file only allows a series of linear motion, but
the framework for curved (i.e. circular) motion is in place. The
following is an example of a real state file (note that the order of the
motions is important):
==============================
StartTime = 2014/09/03 13:37:05.220
#NatureOfMotion, StartX, StartY, EndX, EndY, Duration (s)
LINEAR, 0, 0, 50, 100, 5
LINEAR, 50, 100, -20, 20, 1.5
==============================

When specifing the StartTime, the syntax for "13:37:05.220" implies
1:37pm 5 seconds and 220 milliseconds (the trailing 0 is purely
stylistic and can be left out).