Fixed critical race condition in SensorServer
Fixed a critical bug which was a race condition in the SensorServer for
sending data to the DisplayServer. When running the SensorServer on
insufficiently fast hardware  the measurements would eventually
completely freeze and nothing would be sent to the DisplayServer anymore
(due to a Timestage variable incrementing too much).

This section of the code was completely rewritten into a new function,
WaitForNewMeasurements() which should completely eliminate the problem
as well as make the code much easier to read. The new code does take up
time on the RawDataMutex, but it appears to do so without any noticable
negative impact on the receiving of new raw data.

Other changes:

* Display Server now draws a white background rather than leaving a
transparent one.
* Small fix to RealStateParser's constructor not closing a file in a
timely and explicit manner.