Bugfix/Performance++: SensorServer-->DisplayServer
Improved performance of previous commit by removing the requirement to
iterate through the entire RawData dictionary every time a measurement
is being sent to the DisplayServer, the code now keeps track of a
maximum time stage variable instead.

This has the large benefit of freeing up the RawDataMutex which could
still cause problems on very slow machines, as they could spend all
their time checking the list for new measurements and rarely end up
inserting any new measurements.

Other changes:

* Added a new INI variable "SENSOR_SERVER_MEASUREMENT_WAIT_LAG". This
represents the time to wait for measurements to come in before sending
an update to the display server. Measured in ms
* Removed an unnecessary sleep before sending to DisplayServer, which
caused the initial lag to be two times too long
* Hid the connection semaphore in all applicable net code classes and
replaced with a method, WaitForConnection(). This is more readable and
much less error prone than manually calling WaitOne() and Release(). In
related news this also fixed a newly found bug in DisplayServer which
wouldn't ever release the semaphore.