Implemented KalmanFilter2D w/o tracking
Implemented KalmanFilter2d into InitialEstimator, but it does not do
track association. It is currently only designed for single object
tracks. It will average all possible object positions (hopefully usually
only 1) and use that as a measurement.

Implementing track association is still a large priority, along with
improved trilateration. Speaking of, the next update will return
trilaterated object estimates to the DisplayServer as right now the
Kalman Filter'ed data has replaced trilaterated data as a final result.

Note: MLApp needed to be initiated very early on rather than only when
it's needed, as it takes several seconds to start. That's why it's
initated in Program.cs rather than InitalEstimator.cs.