Trilateration drawn + Customizable DisplayServer
Trilaterated data is now drawn as blue, as it originally was. Estimated
data is kept, but now drawn as purple. As such, yet another list of
objectestimates is now sent from SensorServer to DisplayServer, and the
IEstimator interface requires estimators to keep track of this
additional list now as well.

The InitialEstimator now makes use of the initial measurement in the
Kalman filter by default, by usings its position value as the predicted
initial state.

The INI file now implemenents the following variables (all bools):

* USE_INITIAL_MEASUREMENT_AS_STATE : Turn the above mentioned feature
on/off
* DRAW_REAL : Draws additional green dots denoting the real position of
the tracked objects
* DRAW_ADDITIONAL : Draws additional yellow dots denoting the arbitrary
extra information
* DRAW_ESTIMTAED : Draws additional purple dots denoting the estimated
(i.e. filtered) position of the objects
* DRAW_TRILATERATED : Draws additional blue dots denoting the
trilateration of measurement readings, unfiltered.
* DRAW_MEASUREMENTS : Draws additional small grey dots representing a
visualisation of measurement values from individiual sensors (each will
be directly in front of its sensor)
* DRAW_SENSORS : Draws additional red dots and dotted red field of view
lines representing all of the sensors