Implemented Noisy 2D Trilateration
Implemented the first iteration of noisy 2D trilateration, that attempts
to find an approximate coordinate for where the sum of squared errors is
least to each measurement.

While fully functional, the measurement association is only using a
basic distance threshold approach from the middle point of each
measurement arc, and needs to be improved upon.

Other:
*  Added the following options to the INI:
* TRILATERATOR_NOISY_2D_GROUPING_THRESHOLD, a float that determines how
close measurement arcs can be in order to be considered associable.
* TRILATERATOR_NOISY_2D_GRID_DIVISON, an int that determines how finely
to attempt to find a point of least squared error. This value scales
with O(n^2) so beware of setting it too high. A value in the area of
100-500 seems to work well.
* INITIAL_ESTIMATOR_TRILATERATOR: a string that determines if the new
Noisy or old Noiseless trilateration should be used by the
InitialEstimator