Implemented real-estimate error calculation
The DisplayServer now calculates and stores an error of estimation with
respect to the real state it receives. It will calculate both error in
distance and error in distance squared, and then store the values for
minimum, maximum, and mean in a file named ErrorResults.txt, in the same
directory as image output is provided.

It is currently not designed to handle multiple estimates or real states
and will simply average their positions.